requirements.txt
asgiref==3.5.2
certifi==2022.9.14
charset-normalizer==2.1.1
Django==2.2.10
django-mathfilters==1.0.0
gunicorn==20.0.4
idna==3.4
postgres==4.0
psycopg2-binary==2.8.6
python-dotenv==0.21.0
requests==2.28.1
sqlparse==0.4.2
stripe==4.1.0
urllib3==1.26.12
